release: v2.6.8 — io error path preservation + clippy zero-warning
Bug fixes
- B1: find_workspace_root walked past /, /tmp, $TMPDIR, $HOME and picked
  stray ancestor Cargo.toml as workspace root. Both copies (main.rs,
  resolve/references/rust.rs) now stop at boundary dirs and recognise
  pyproject.toml, package.json, go.mod, pom.xml, build.gradle, setup.py.
- B2: should_skip_entry excludes 14 more build/cache dirs (dist, build,
  __pycache__, .venv, venv, .pytest_cache, .mypy_cache, .tox, .next,
  .nuxt, .cache, .gradle, .idea, .vscode) from preview workspace clone.
- B3: new src/io_ext.rs provides path-preserving read/read_to_string/
  write helpers. All bare `?` propagation of std::fs calls migrated.
- B4: validate_with_cargo cargo spawn failure now surfaces project_dir
  instead of <unknown>.
- B5: patch::validate_utf8_span takes file_path: &Path (breaking, no
  internal callers); InvalidSpan errors now name the real file.
- symbol::parser_for_language takes file_path: &Path (breaking, 3
  internal callers updated); Parse errors now name the real file.

Catch-all removal
- Removed `From<std::io::Error> for SpliceError`. The impl produced
  Io { path: "<unknown>", .. } for any bare `?`, leaking the
  placeholder everywhere. Six remaining bare-? sites migrated to
  explicit map_err with originating path. Future bare `?` on io::Error
  in code returning Result<_, SpliceError> no longer compiles.

Dead code
- Removed unused BatchResult.spec_path field (never read; comment
  "Set by caller" was inaccurate). Breaking for hypothetical external
  Rust consumers.
- Deleted dead test helper functions/methods across 14 test files.

Clippy: 338 → 0
- cargo clippy --all-targets -- -D warnings clean from a cold rebuild.
- Mix of cargo clippy --fix, manual fixes (sort_by_key, regex hoisted
  out of loops, &PathBuf → &Path, identical-block dedup, doc list
  indentation, Box<Relationships> for the large Err variant), and
  project-rule-compliant #[allow(..., reason = "...")] for CLI handler
  arg counts and Arc<MagellanIntegration> shared ownership.

Tests
- Added tests/find_workspace_root_boundary.rs (2 integration tests).
- Added tests/io_error_path_context.rs (2 integration tests).
- All 1453 tests pass (12 ignored, 0 failed).

+++ src/ingest/rust.rs
@@ -119,12 +119,13 @@ fn extract_symbols(
 /// In tree-sitter Rust grammar:
 /// - Inherent impl: `impl StructName` -> has `type:` field pointing to StructName
 /// - Trait impl: `impl Trait for StructName` -> has `trait:` field (Trait) and `type:` field (StructName)
+///
 /// The `type:` field ALWAYS contains the struct name being implemented.
 fn extract_impl_name(node: &tree_sitter::Node, source: &[u8]) -> Option<String> {
     // Access the 'type' field which always contains the struct name
     let type_node = node.child_by_field_name("type")?;
 
-    let name_bytes = &source[type_node.start_byte() as usize..type_node.end_byte() as usize];
+    let name_bytes = &source[type_node.start_byte()..type_node.end_byte()];
     std::str::from_utf8(name_bytes).ok().map(|s| s.to_string())
 }
